initial code commit
This is still a little buggy (with bugs, not a method of conveyance). I
think some software improvement with the rotor search algorithm (like by
adding any subtlety) will fix a lot of the remaining issues.
nevertheless, here it is. As will all of these projects, a circuit
diagram would make a world of difference in terms of actually using it
outside of  our lab.