wow, it has taken a bit of work to get here
The zero function seems to find zero every time, but hasn’t been bomb
proofed against starts (there are some unprotected seek functions in
there).
Added EEEprom logging of the position of the arm. This may or may not
be useful, as at present, it is expected to zero itself, and not really
trust the counter.
added a little serial reader to allow testing of bottle targeting
without hitting too many buttons.
some functions in here are no longer used. i really did have to frig
around a bunch to get the arm zeroing function to work properly.

added a timer to keep track of average seek time, allowing the program
to try to go past false positives for finding the next spot.
added a retry section in the seek forward/reverse to try to get past a
sticky point. I think that needs to be improved.